pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hermes-bridge"
version = "0.1.0"
description = "Mobile-ready HTTP bridge for hermes-agent."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "hermes-bridge contributors" }]
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "pydantic>=2.7",
    "pydantic-settings>=2.3",
    "sse-starlette>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
    "mypy>=1.10",
    "respx>=0.21",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "TID251"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"hermes".msg = "Do not import from hermes internals; use HTTP/CLI/filesystem."
"hermes_cli".msg = "Do not import from hermes internals."
"gateway".msg = "Do not import from hermes internals."
"agent".msg = "Do not import from hermes internals."

[tool.mypy]
python_version = "3.12"
strict = true
packages = ["hermes_bridge"]
mypy_path = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
